+++ src/mcp.c
@@ -814,7 +814,7 @@ mcp_frame_output_mesg(McpFrame * mfr, McpMesg * msg)
 		strcat(out, MCP_ARG_DELIMITER);
 		strcat(out, MCP_SEPARATOR);
 		out += strlen(out);
-		sprintf(datatag, "%.8lX", random() ^ random());
+		sprintf(datatag, "%.8lX", (unsigned long)(random() ^ random()));
 		strcat(out, datatag);
 	}
 
@@ -1303,7 +1303,7 @@ mcp_basic_handler(McpFrame * mfr, McpMesg * mesg, void *dummy)
 			mcp_mesg_init(&reply, MCP_INIT_PKG, "");
 			mcp_mesg_arg_append(&reply, "version", "2.1");
 			mcp_mesg_arg_append(&reply, "to", "2.1");
-			sprintf(authval, "%.8lX", random() ^ random());
+			sprintf(authval, "%.8lX", (unsigned long)(random() ^ random()));
 			mcp_mesg_arg_append(&reply, "authentication-key", authval);
 			mfr->authkey = (char *) malloc(strlen(authval) + 1);
 			strcpy(mfr->authkey, authval);
@@ -1731,6 +1731,9 @@ mcp_internal_parse(McpFrame * mfr, const char *in)
 
 /*
 * $Log: mcp.c,v $
+* Revision 1.17  2002/03/17 07:55:25  winged
+* Fixing pedantic warnings
+*
 * Revision 1.16  2002/02/26 06:58:32  revar
 * Fixed bug in MCP where _datatag wasn't being preceeded by a separator space.
 *
